Fixed errors in p1 and p2
Found and fixed error with concatanation in p1.There were being a conflict between the use of pointers. Strings should be copied before work with them as they are set as pointers and they get directly affected.

Fixed error regarding trash when identifying an ID_ARITM token. It was
due to the size of yyleng, it doesn't take into account the additional
space for the trailing character in C (\0)."